+++ src/data_objects/WebPage.m
@@ -51,17 +51,17 @@
     return c;
 }
 
-- (int) compareTo:(WebPage *)ob
+- (long) compareTo:(WebPage *)ob
 {
 	return (ob.web_page_id == self.web_page_id);
 }
 
 - (NSString *) description
 {
-    return [NSString stringWithFormat:@"WebPage- Id:%d\tName:%@\tURL:%@\t",self.web_page_id,self.name,self.url];
+    return [NSString stringWithFormat:@"WebPage- Id:%ld\tName:%@\tURL:%@\t",self.web_page_id,self.name,self.url];
 }
 
-- (int) icon_media_id
+- (long) icon_media_id
 {
     if(!icon_media_id) return DEFAULT_WEB_PAGE_ICON_MEDIA_ID;
     return icon_media_id;
